+++ lib/Entry.js
@@ -22,7 +22,6 @@ import defaultTheme from './themes/default.js'
 
 import Manifest from './Manifest.js'
 import Module from './Module.js'
-// import CoreModule from './modules/core.js'
 
 import ImportRule from './atrules/import/ImportRule.js'
 import FileResource from './atrules/import/FileResource.js'
@@ -40,8 +39,6 @@ import Theme from './atrules/theme/Theme.js'
 import Stylesheet from './Stylesheet.js'
 import Partial from './Partial.js'
 
-// import FileUtils from './utilities/FileUtils.js'
-// import ImportUtils from './utilities/ImportUtils.js'
 import CSSUtils from './utilities/CSSUtils.js'
 import QueueUtils from './utilities/QueueUtils.js'
 
@@ -76,7 +73,6 @@ export default class Entry extends Stylesheet {
   }
 
   render (cb) {
-    console.log(this.#manifest.json);
     let results = []
 
     QueueUtils.run({
@@ -112,12 +108,12 @@ export default class Entry extends Stylesheet {
             componentResets(annotations, this.#manifest.components),
             components(annotations, this.#manifest.components, theme.components),
 
-            // typography(annotations),
             // atrules,
             // hoist,
             functions({ functions: CONFIG.functions }),
             // ...this.#getInternalModules(theme),
             env(CONFIG.env),
+            typography(annotations),
             namespace,
             ...post
           ]).process(this.root.clone(), {
@@ -178,18 +174,20 @@ export default class Entry extends Stylesheet {
 
     this.walkAtRules('component', atrule => {
       let componentRule = new ComponentRule(atrule)
-      (componentRule.superclass ? extensions : components).push(componentRule)
+
+      if (componentRule.superclass) {
+        extensions.push(componentRule)
+      } else {
+        components.push(componentRule)
+      }
+
       componentRule.remove()
     })
 
     components.forEach(component => this.#manifest.addComponent(new Component(component)))
 
     extensions.sort((a, b) => {
-      if (extensions.some(extension => extension.name === a.superclass)) {
-        return 1
-      }
-
-      return 0
+      return extensions.some(extension => extension.name === a.superclass) ? 1 : 0
     }).forEach(extension => {
       let { superclass } = extension
       let parent = this.#manifest.getComponent(superclass)
