+++ lib/atrules/apply/Application.js
@@ -9,34 +9,23 @@ export default class Application {
     this.#source = applyRule
   }
 
+  get bounds () {
+    return this.#source.bounds
+  }
+
   get margin () {
     let { margin } = this.#source
-
-    if (!!margin) {
-      return new Margin(margin)
-    }
-
-    return null
+    return !!margin ? new Margin(margin) : null
   }
 
   get padding () {
     let { padding } = this.#source
-
-    if (!!padding) {
-      return new Padding(padding)
-    }
-
-    return null
+    return !!padding ? new Padding(padding) : null
   }
 
   get typeset () {
     let { typeset } = this.#source
-
-    if (!!typeset) {
-      return new Typeset(typeset)
-    }
-
-    return null
+    return !!typeset ? new Typeset(typeset) : null
   }
 }
