+++ lib/atrules/apply/ApplyRule.js
@@ -1,22 +1,24 @@
 import AtRule from '../AtRule.js'
 import MarginRule from './MarginRule.js'
 import PaddingRule from './PaddingRule.js'
-// import Display from './Display.js'
 import TypesetRule from './TypesetRule.js'
-// import Typeset from './Typeset.js'
+import MediaQueryRule from '../media/MediaQueryRule.js'
 
 // import CSSUtils from '../../utilities/CSSUtils.js'
 // import QueueUtils from '../../utilities/QueueUtils.js'
 // import SelectorUtils from '../../utilities/SelectorUtils.js'
 
 export default class ApplyRule extends AtRule {
+  #bounds
   #typesetRule
   #marginRule
   #paddingRule
   #isShorthand
+  #source
 
   constructor (applyRule) {
     super(applyRule)
+    this.#source = applyRule
     this.#isShorthand = !(applyRule.nodes?.length > 0)
 
     applyRule.walkAtRules(atrule => {
@@ -34,6 +36,12 @@ export default class ApplyRule extends AtRule {
           break
       }
     })
+
+    this.#bounds = this.#getBounds()
+  }
+
+  get bounds () {
+    return this.#bounds
   }
 
   get isShorthand () {
@@ -67,6 +75,22 @@ export default class ApplyRule extends AtRule {
     }
 
     return this.#typesetRule ?? null
+  }
+
+  #getBounds = () => {
+    let { parent } = this.#source
+
+    if (parent.type === 'rule') {
+      parent = parent.parent
+    }
+
+    if (parent.type === 'atrule' && parent.name === 'media') {
+      let query = new MediaQueryRule(parent)
+      let { min, max } = query.width
+      return (!!min || !!max) ? query.width : null
+    }
+
+    return null
   }
 }
