+++ lib/atrules/apply/Display.js
@@ -14,7 +14,7 @@ export default class Display {
   constructor (atrule) {
     Array.isArray(atrule) ? this.#processShorthand(...atrule) : this.#processLonghand(atrule)
 
-    if (/*!!atrule.display && */['x', 'y', 'top', 'right', 'bottom', 'left'].every(attr => !atrule[attr])) {
+    if (['x', 'y', 'top', 'right', 'bottom', 'left'].every(attr => !atrule[attr])) {
       this.#properties.x = true
       this.#properties.y = true
     }
