+++ lib/atrules/apply/Typeset.js
@@ -1,5 +1,4 @@
 export default class Typeset {
-  // sizeSet = false
   #size = 0
   #relative = false
