+++ lib/plugins/typography.js
@@ -18,6 +18,7 @@ export default postcss.plugin('chassis-typography', annotations => {
     })
 
     // applications.forEach(application => {
+    //   console.log('BOUNDS: ', application.bounds);
     //   console.log(`MARGIN: `, application.margin?.properties ?? null)
     //   console.log(`PADDING: `, application.padding?.properties ?? null)
     //   console.log(`TYPESET: `, application.typeset?.size ?? null, application.typeset?.relative ?? null)
@@ -25,7 +26,7 @@ export default postcss.plugin('chassis-typography', annotations => {
     // })
 
     let { viewports } = CONFIG
-    console.log(viewports)
+    // console.log(viewports)
 
     resolve(root)
   })
