+++ lib/themes/default.js
@@ -49,12 +49,7 @@ const raw = parser.parse(`
       button {
         background: lightgrey;
         border-radius: 3px;
-
-        @apply {
-          @padding {
-            display: inline-block;
-          }
-        }
+        @apply padding;
       }
 
       table {
