+++ lib/utilities/SelectorUtils.js
@@ -1,6 +1,6 @@
 import nesting from 'postcss-nesting'
 
-import MediaQuery from '../atrules/media/MediaQuery.js'
+import MediaQueryRule from '../atrules/media/MediaQueryRule.js'
 import CSSUtils from './CSSUtils.js'
 
 export default class SelectorUtils {
@@ -74,7 +74,6 @@ export default class SelectorUtils {
         let rule = CSSUtils.createRule(node.selector)
         previous.append(rule)
         previous = rule
-
         return
       }
