+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@chassis/core",
   "type": "module",
-  "version": "3.0.0-alpha.21",
+  "version": "3.1.0-alpha.1",
   "description": "CSS4 pre-processor and responsive framework for modern UI development",
   "main": "index.js",
   "scripts": {
